lua: ensure that host-lua does not depend on host-readline and host-ncurses
Commit ad6af5819f88bac9ca7f50e565c4fc68d3e408cc removed the
HOST_LUA_DEPENDENCIES assignement to the empty variable, so from that
commit, the dependencies of host-lua are automatically computed from
the dependencies of lua. Since lua can depend on readline and ncurses,
it means that host-lua can depend on host-readline and host-ncurses,
and the host-readline package does not exist in Buildroot. Since there
is no reason to build host-lua with readline/ncurses support, just
re-add the HOST_LUA_DEPENDENCIES variable.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Cc: Francois Perrad <fperrad@gmail.com>