xstroke: fix link issue
The allyespackageconfig builds have trigerred a link issue on
xstroke. In fact, xstroke must be linked against Xrender, Xext and dl
in addition to the other libraries it was already be linked
against. We fix that by adding a patch that modifies configure.ac and
Makefile.am, and enabling autoreconf for this package.

In addition, since xstroke uses directly Xrender and Xext, we add
those dependencies to the package. In practice, it doesn't change
anything, since some of the xstroke dependencies were already pulling
those libraries, but as xstroke uses directly Xrender and Xext, it
makes sense to have them as dependencies as well.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>