firmwares: add a new package category to locate all firmwares
We have quite a few packages that are dedicated to providing firmwares
for miscellanous devices. Those package are spread out in the big
"Hradware handling" menu, so are a bit hard to find.

Move all those packages to their own sub-menu:
    Package Selection for the target  --->
        Hardware handling  --->
            Firmwares  --->
                [ ] b43-firware
                [ ] linux-firmware
                [ ] ux500-firmware

Hopefully, this will make it nicer, when new firmware-providing packages
are added in the future (eg. in the pipe: firmware for the RPI GPU).

Fix a typo in the 'b43-firmware' prompt (missing 'm').

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Acked-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>