xserver_xorg-server: option to build KDrive input drivers
When the Kdrive variant of the X server is selected, the drivers are
directly built into the X server. The X server therefore provides
options to enable or disable certain drivers, especially input
drivers.

This patch adds options to be able to enable or disable the evdev, kbd
and mouse drivers of Kdrive.

[Peter: fix Config.in indentation as pointed out by Arnout]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>