linux, uboot, mxs-bootlets, barebox: avoid double slash in CUSTOM_TARBALL
Closes #5846

The $(dir ...) function leaves a slash at the end, so that the LINUX_SITE
variable for a custom tarball ends in a slash. The DOWNLOAD macro adds
another slash between SITE and SOURCE, which results in a double slash in
the download URL.

Fix this by stripping off the final slash from the _SITE in all packages that
have a custom tarball.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>