libffi: switch to using a Git version, brings Microblaze and Xtensa support
We already had backported patches for the Blackfin and AArch64
support, and now we would have needed to also backport the Xtensa
support and the Microblaze support. This starts to get crazy.

Let's switch to using a Git version from Github, until libffi finally
releases a new stable version.

In order to achieve this, we also need to:

 * autoreconf the package, so that a libffi.pc file gets generated

 * manually install the libffi.pc file, because it doesn't get
   installed by libffi Makefile, for some reason

 * remove the part of the target post install hook that was messing
   with libffi.pc, since it was modifying the one in STAGING_DIR
   (which is odd for a target post install hook), and the libffi.pc
   file is anyway not installed to the target anymore.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>