linknx: forcilbly disable smtp support
linknx has support for SMTP via libesmtp, however there's no clean way
to specify where libesmtp-config lives.
So when libesmtp is built it finds it via the headers and is
automatically enabled, but it fails to get the appropiate link line from
libesmtp-config because it's not found.
Fix it by forcibly disabling it. If someone is interested in working
with upstream about this an option like --with-libcurl should be
implemented in configure.
Solves:
http://autobuild.buildroot.net/results/90ebb383d7a2d7bdf09ffac65e2504da7b6d19f5/

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>