ne10: new package
The Ne10 project has been set up to provide a set of common, useful
functions which have been heavily optimized for the ARM Architecture
and provide consistent well tested behavior that can be easily
incorporated into applications. C interfaces to the functions are
provided for both assembler and NEON implementations.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>