make: undeprecate package as it is used by the uClibc test suite
The uClibc test suite can be installed on the target thanks to the
BR2_UCLIBC_INSTALL_TEST_SUITE option, but it requires make to be
installed on the target. Since the make package is currently marked as
deprecated, it leads to the following warning:

warning: (BR2_UCLIBC_INSTALL_TEST_SUITE) selects BR2_PACKAGE_MAKE which has unmet direct dependencies (BR2_DEPRECATED)

We fix this by undeprecating the make package, and also update its
description to explain why we have the make package even if we don't
have support for building a toolchain for the target.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>