Switch to using WebKitGTK+ 2.3.5
The plan is to use WebKitGTK+ 2.3.5.

The package should be auto-recofigured.

libcurl is removed from the list of dependencies.

ar flags are set to 'cru' since thin archives are not supported by the
buildroot.

Additional features that require extra dependencies are for now disabled --
credential storage, geolocation, media support (HTML5 video, Web Audio API).
At least HTML5 video suuport will be enabled later on, when the basic
functionality is ensured.

Enable the X11 target when building with XORG7 package enabled, and enable the
Wayland target when building the Wayland package.