Add three necessary patches that make 2.3.5 buildable.
avoid-duplicated-symbols.patch -- avoids duplicated symbols in WebCore plugins
code when building only for the Wayland target.

ax-compare-versions-macro.patch -- include a helper macro that compares two
strings representing versions.

remove-gtk2-dependency.patch -- Avoid the additional libgtk2 dependency and
link the WebKitPluginProcess against libgtk3. GTK+ 2 is used to make it
possible to use the Flash plugin, but the latter is not available on the RPi
anyways.