Avoid stdc-predef.h inclusions when compiling extract-strings
The $(CC_FOR_BUILD) compiler (for instance GCC 4.7 and newer) can implicitly
include stdc-predef.h header, which isn't provided by eglibc 2.12, so it falls
back to using the base system's header under /usr/include. This can cause build
failures when the stdc-predef.h header continues to include the <bits/predefs.h>
header which is provided by eglibc 2.12, but is not compatible with the header
that's provided by the system's eglibc that's of more recent version, e.g. 2.17.

Pass -ffreestanding to the $(CC_FOR_BUILD) compiler to completely avoid the
inclusion of stdc-predef.h.